+++ 1st-gen/packages/menu/src/MenuItem.ts
@@ -331,7 +331,6 @@ export class MenuItem extends LikeAnchor(
         if (this.disabled) {
             event.preventDefault();
             event.stopImmediatePropagation();
-            event.stopPropagation();
             return false;
         }