+++ 2nd-gen/packages/core/shared/base/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2025 Adobe. All rights reserved.
+ * Copyright 2026 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -15,12 +15,12 @@ export {
     SpectrumMixin,
     type SpectrumInterface,
 } from './Base.js';
+export { defineElement } from './define-element.js';
 export {
-    SizedMixin,
-    type ElementSize,
     ElementSizes,
+    SizedMixin,
     type DefaultElementSize,
+    type ElementSize,
     type SizedElementInterface,
 } from './sizedMixin.js';
-export { defineElement } from './define-element.js';
 export { version } from './version.js';